converting price to int

+++ apiapp/models.py
@@ -1,7 +1,7 @@
 import uuid
 from django.db import models
 from django.db.models.fields import CharField, AutoField, BooleanField
-from django.forms import DecimalField
+from django.forms import DecimalField, IntegerField
 
 # Create your models here.
 
@@ -23,7 +23,7 @@ class Coffee(models.Model):
     description = models.TextField(max_length=512)
     imageurl = models.TextField(max_length=128, null=True)
     slug = models.fields.SlugField
-    price = DecimalField(min_value=0)
+    price = IntegerField(min_value=0)
     def __str__(self) -> str:
         return self.name
